requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.32.0,<3.0.0
PyMuPDF>=1.24.0,<2.0.0
pydantic>=2.8.0,<3.0.0

